Add files via upload
A python WebCrawler that crawls a random wikipedia page and open the first link. Then on that page it clicks the first link in the main body of the article text and it keeps going. There is a common saying that if you kept click on the first link you will land to the 'Philosophy' page as it's the mother of all descendants topics. This script tries to automate it using web crawling.