Add files via upload
A project using of Python to explore data related to bikeshare systems for three major bikeshare systems in the United States. Performing data wrangling to unify the format of data from the three systems and writing code to compute descriptive statistics. Also making use of a package that is not part of the standard Python library to help  visualize the data.

This done using Jupyter notebook.